requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-autofixture==0.12.1
django-jalali-date==0.2.10
idna==2.8
influxdb==5.2.2
jdatetime==3.4.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
